// reelease-ai-next/src/types/api.ts
export interface ApiError {
  data?: {
    message?: string
    error?: string
    errors?: string[]
  }
  status?: number
  message?: string
}